+++ wms/patches/po_investment_status_migration.py
@@ -1,0 +1,25 @@
+import frappe
+
+
+def execute():
+	submitted_to_po_status_list = frappe.get_all(
+		"WMS PO Investment", filters={"status": "Submitted to PO"}, fields=["name"]
+	)
+	[
+		frappe.set_value("WMS PO Investment", item["name"], "status", "Submitted")
+		for item in submitted_to_po_status_list
+	]
+
+	passbook_received_status_list = frappe.get_all(
+		"WMS PO Investment", filters={"status": "Passbook Received"}, fields=["name"]
+	)
+	for item in passbook_received_status_list:
+		frappe.set_value("WMS PO Investment", item["name"], "status", "Active")
+		frappe.set_value("WMS PO Investment", item["name"], "passbook_status", "With Us")
+
+	passbook_sent_to_customer_status_list = frappe.get_all(
+		"WMS PO Investment", filters={"status": "Passbook Sent to Customer"}, fields=["name"]
+	)
+	for item in passbook_sent_to_customer_status_list:
+		frappe.set_value("WMS PO Investment", item["name"], "status", "Active")
+		frappe.set_value("WMS PO Investment", item["name"], "passbook_status", "With Client")